refactor: consolidate component imports and export components from their respective directories
